add step def base page
