--- src/main/java/ru/sbtqa/tag/pagefactory/stepdefs/AbstractGenericStepDefs.java
@@ -1,11 +1,0 @@
-package ru.sbtqa.tag.pagefactory.stepdefs;
-
-public class AbstractGenericStepDefs {
-
-    public String getLanguage() {
-        String[] packages = this.getClass().getCanonicalName().split("\\.");
-        String currentLanguage = packages[packages.length - 2];
-        
-        return currentLanguage;
-    }
-}
